requirements.txt
python-telegram-bot>=20.1
openai>=0.28.1
tiktoken>=0.3.0
PyYAML>=6.0
pymongo>=4.3.3
python-dotenv>=0.21.0
